1_The_Physical/Level_1_Stateless_Guess_Point/1_Arrange_Act_Assert_Backwards/2_1_Password_Validator/src/index.ts
interface IResult {
  result: boolean;
  errors: string[];
}

export default function checkPassword(input: string): IResult {
  const noOfDigits = input
    .split("")
    .filter((char) => typeof char === "number").length;
  const isValidLength = input.length >= 5 && input.length <= 15;

  if (!isValidLength) {
    return {
      result: false,
      errors: ["InvalidLengthError"],
    };
  }

  if (noOfDigits === 0) {
    return {
      result: false,
      errors: ["NoDigitError"],
    };
  }

  return {
    result: true,
    errors: [],
  };
}
